Im tired

+++ PHP/User.delete.php
@@ -6,25 +6,40 @@
   $conn = CConexion::ConexionBD();
   
   $rol=$_SESSION['ROL'];
-  $path= 'image/'.$rol;
+  $path= 'image/users/'.$rol;
   $img=$_SESSION['img'];
   
   $delete_query=<<<SQL
   DELETE FROM cuenta
-    WHERE "ROL"=$1
+    WHERE rol=$1
   SQL;
 
-  unlink($img);
-  rmdir($path);
-
+  if(!empty($img)){
+    unlink($img);
+    rmdir($path);
+  }
   $delete=pg_prepare($conn,"Delete",$delete_query);
   $delete=pg_execute($conn,"Delete",array($rol));
 
-  session_destroy();
-  echo'
-  <script language="javascript">
-        alert("Cuenta eliminada Correctamente");
-        window.location = "./index.php";
-  </script>
-  ';
+
+
+  if($delete){
+    $error=pg_result_error($delete);
+    echo $error;
+    session_destroy();
+    echo('
+          <script language="javascript">
+              alert("wea borrada");
+              //window.location = "./UserSignIn.php";  
+          </script>
+      ');
+  }else{
+      
+      echo('
+          <script language="javascript">
+              alert("wea no borrada");
+              //window.location = "./UserSignIn.php";  
+          </script>
+      ');
+  }
 ?>
